requirements.txt
tensorflow-gpu
imgaug